pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dalle-mini-tools"
version = "0.1.0"
description = "A collection of tools for generating dalle-mini images"
authors = ["Mark Kockerbeck","Matt Mollison"]

[tool.poetry.dependencies]
dalle-mini = "^0.1.1"
fire = "^0.4.0"
Flask = "^2.1.2"
ipywidgets = "^7.7.1"
pySqsListener = "^0.8.10"
python = ">=3.10,<3.11"
python-slugify = "^6.1.2"
tokenizers = "~=0.11.6"
vqgan-jax = {git = "https://github.com/patil-suraj/vqgan-jax.git", rev = "main"}
slack-sdk = "^3.17.2"
slack-bolt = "^1.14.0"
tqdm = "^4.64.0"
jaxlib = [
    {platform = 'linux', markers = "platform_machine == 'x86_64'", url = "https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.3.10+cuda11.cudnn82-cp310-none-manylinux2014_x86_64.whl"},
    {platform = 'darwin', markers = "platform_machine == 'arm64'", url = "https://storage.googleapis.com/jax-releases/mac/jaxlib-0.3.10-cp310-none-macosx_11_0_arm64.whl"},
 ]
aiohttp = ">=0.0.0"

[tool.poetry.dev-dependencies]
black = "~22.3.0"
flake8 = "~4.0.1"
ipdb = "^0.13.9"
isort = "~5.10.1"
mypy = "~0.960"
# pytest = "~7.1.2"

[tool.black]
line-length = 88
preview = true

[tool.isort]
profile = "black"
